el pepe2

+++ vista_cliente/administrar_cita/agendar/eliminar_cita.php
@@ -1,18 +1,20 @@
 <?php
-include('../../../connection/connection.php');  // Asegúrate de que la ruta a conexion.php es correcta
+session_start();
+include('../../../connection/connection.php');
 
-$id_cita = $_POST['id_cita'];
+if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+    $id_cita = $_POST['id_cita'];
 
-$stmt = $conn->prepare("DELETE FROM citas WHERE id_cita = ?");
-$stmt->bind_param("i", $id_cita);
+    $stmt = $conn->prepare("DELETE FROM citas WHERE id_cita = ?");
+    $stmt->bind_param("i", $id_cita);
 
-if ($stmt->execute()) {
-    header("Location: ../index.php"); // Redirige de vuelta a la página principal después de la eliminación
-    exit();
-} else {
-    echo "Error: " . $stmt->error;
+    if ($stmt->execute()) {
+        header("Location: ./index.php");
+    } else {
+        echo "Error: " . $stmt->error;
+    }
+
+    $stmt->close();
+    $conn->close();
 }
-
-$stmt->close();
-$conn->close();
 ?>
